+++ src/eth/EthBalances.js
@@ -13,7 +13,7 @@ import {i18nValidator, i18nBalance} from '../i18n';
 export const EthBalances = ({addresses, fetch}) => {
   if (!addresses) return null;
   const [loadEth, balanceEth] = useEthscanBalance(addresses);
-  const [loadToken, balanceToken] = useEthscanTokensBalance(addresses, fetch);
+  const [loadToken, balanceToken] = useEthscanTokensBalance(fetch, addresses);
 
   if (loadEth && loadToken) {
     return (<Text>{t('query', {i18n: i18nBalance})}</Text>)